LPD-82593 Moving also DBResourceProvider to portal-kernel
